+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Adopt Spell Engine 1.10 changes
 - NeoForge version no longer depends on Forgified Fabric API
 - Fully translated content, now supporting 20 languages
+- Removed Accessories mod support
+- Added Curious API slot mod support (NeoForge only)
 
 # 1.3.0